# Statically choose lowest (least lossy) compression for JPEG and TIFF