@alexpevzner 2025-06-26
Hi!

I will accept the choice of the lowest possible compression level with the following small modification: it should not be applied to PNG and BMP formats.

With JPEG it will most likely increase the image quality, and may also increase the image quality with TIFF, because TIFF may be just a wrapped JPEG (we can't know a priori). But with BMP it will do just nothing and with PNG it will only increase image processing time at the scanner side.

Regarding changes in image formats prioritization, I will not accept it. The rationale is following.

1. BMP. This format is **very** rarely used by scanners. I've seen it only once in wild, and that device didn't offer any other format, so adding BMP was a matter to enable scanning from this device.

BMP may be as simple as just a matrix of pixels with small headers, or may be pelleted, may wrap JPEG and, as far as I remember, may use its own compression. `sane-airscan` doesn't support all BMP variants, only the simplest one. This OK when BMP is used only as a last resort for those few devices that really require it, but not OK, if in result of the format prioritization change it will be used more broadly, as there is a chance that there will be some devices that will use more complex format of BMP that `sane-airscan` really support, so instead of increased quality users of these devices will see the failed scans. I don't want to take such a risk.

Note also, that uncompressed BMP is very large. A4 RGB picture at 1200 DPI takes about 1/2 Gb of RAM. The `sane-airscan` attempts to minimize the memory consumption, by keeping only 1 row of the uncompressed image in memory. With BMP it obviously doesn't work.

2. TIFF. This format is **very** complex. And it doesn't guarantee that it can be uncompressed in the line-by-line mode. Also, being using it as a last resort format, I'm not forced to implement it at full. Only some small subset of devices have to fall back to TIFF, and probably my implementation is good enough for them (at least, there is no compatibility problems known so far). If I will place TIFF before JPEG, much more devices will use it and I predict a lot of compatibility problems and noticeable regression.

Please notice also, that this is not uncommon for devices to use TIFF (and PDF) just as a container for JPEG. I guess many of them use hardware JPEG encoder, and image is compressed before it is placed into the device memory, so allowing high-resolution scanning for devices with the limited amount of RAM. So if internally device works with JPEG images only, changing the container format will not improve the image quality.